get subgraph definitions in compose mod
